// src/app/models/aporte-afectado.ts
export interface AporteAfectado{
    nroAporte: string;
    fechaAporte: Date;
    opcion: string;
    sbota: string;
    valorAporte: number;
    saldoAporte: number;
    saldoRendimientos: number;
    liberaCntge: string;
    rfteEstimadaCapital: number;
    comiRetiroAnticipado: number;
    valorDisponible: number;
    saldoRetencion: number;
    valorEgreso: number;
    valorTotalEgreso: number;
    feteFuenteEstimada: number;
    costosOperacion: number;
    rendimientosARetirar: number;
    reteFuenteCapital: number;
    reteFuenteAPracticar: number;
    contingenteAPracticar: number;
    comiRetAntiA: number;
}